requirements-dev.txt
# Python dev tools — install with: pip install -r requirements-dev.txt
ruff>=0.4.0
pyright>=1.1.350
scrapegraphai>=1.13
requests>=2.31
pydantic>=2.5
python-dotenv>=1.0
jinja2>=3.1
weasyprint>=61
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
pytest>=8.0
pytest-mock>=3.12
pytest-cov>=5.0
langsmith
fastapi
uvicorn

